Minor improvements. Copy some arrays and strings instead of retaining them, try to prevent bugs before they happen. Quiet more 64-to-32 integer warnings.

+++ XADMaster/XAD7ZipBranchHandles.m
@@ -87,30 +87,30 @@
 	x86_Convert_Init(state);
 }
 -(int)decodeBlock:(uint8_t *)block length:(int)length offset:(off_t)pos
-{ return x86_Convert(block,length,(UInt32)pos,(UInt32 *)&state,0); }
+{ return (int)x86_Convert(block,length,(UInt32)pos,(UInt32 *)&state,0); }
 @end
 
 @implementation XAD7ZipPPCHandle
 -(int)decodeBlock:(uint8_t *)block length:(int)length offset:(off_t)pos
-{ return PPC_Convert(block,length,(UInt32)pos,0); }
+{ return (int)PPC_Convert(block,length,(UInt32)pos,0); }
 @end
 
 @implementation XAD7ZipIA64Handle
 -(int)decodeBlock:(uint8_t *)block length:(int)length offset:(off_t)pos
-{ return IA64_Convert(block,length,(UInt32)pos,0); }
+{ return (int)IA64_Convert(block,length,(UInt32)pos,0); }
 @end
 
 @implementation XAD7ZipARMHandle
 -(int)decodeBlock:(uint8_t *)block length:(int)length offset:(off_t)pos
-{ return ARM_Convert(block,length,(UInt32)pos,0); }
+{ return (int)ARM_Convert(block,length,(UInt32)pos,0); }
 @end
 
 @implementation XAD7ZipThumbHandle
 -(int)decodeBlock:(uint8_t *)block length:(int)length offset:(off_t)pos
-{ return ARMT_Convert(block,length,(UInt32)pos,0); }
+{ return (int)ARMT_Convert(block,length,(UInt32)pos,0); }
 @end
 
 @implementation XAD7ZipSPARCHandle
 -(int)decodeBlock:(uint8_t *)block length:(int)length offset:(off_t)pos
-{ return SPARC_Convert(block,length,(UInt32)pos,0); }
+{ return (int)SPARC_Convert(block,length,(UInt32)pos,0); }
 @end
